% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{addCreator}
\alias{addCreator}
\title{Add another author to the meta data of the file.}
\usage{
addCreator(wb, Creator)
}
\arguments{
\item{wb}{A workbook object}

\item{Creator}{A string object with the name of the creator}
}
\description{
Just a wrapper of wb$addCreator()
}
\examples{

wb <- createWorkbook()
addCreator(wb, "test")
}
\author{
Philipp Schauberger
}
