% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addFilter}
\alias{addFilter}
\title{add filters to columns}
\usage{
addFilter(wb, sheet, rows, cols)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{cols}{columns to add filter to.}

\item{rows}{A row numbers}
}
\description{
add filters to columns
}
\details{
adds filters to worksheet columns, same as filter parameters in writeData.
writeDataTable automatically adds filters to first row of a table.
}
\examples{
##2014 April 21st to 25th
x <- c(41821.8127314815, 41820.8127314815)
convertToDateTime(x)
}
\seealso{
\code{\link{writeData}}
}

