% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{all.equal}
\alias{all.equal}
\alias{all.equal.Workbook}
\title{Check equality of workbooks}
\usage{
\method{all.equal}{Workbook}(x, y)
}
\arguments{
\item{x}{A \code{Workbook} object}

\item{y}{A \code{Workbook} object}
}
\description{
Check equality of workbooks
}

