% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as_POSIXct_utc}
\alias{as_POSIXct_utc}
\title{Convert to POSIXct with timezone UTC}
\usage{
as_POSIXct_utc(x)
}
\arguments{
\item{x}{something as.POSIXct can convert}
}
\description{
Convert to POSIXct with timezone UTC
}
\keyword{internal}
