% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openXL.R
\name{chooseExcelApp}
\alias{chooseExcelApp}
\title{Choose excel app}
\usage{
chooseExcelApp()
}
\description{
First the global option for openxlsx.excelApp is checked.  If this is null, 
  an app is attempted to be found.  If the session is interactive and 
  multiple apps are found, the user will have the opportunity to select the 
  appropriate app; otherwise a warning is thrown and the first found will
  be chosen.
This does not set global options
}
\details{
Tries to find the appropriate excel app to use with openXL
}
