% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{convertFromExcelRef}
\alias{convertFromExcelRef}
\title{Convert excel column name to integer index}
\usage{
convertFromExcelRef(col)
}
\arguments{
\item{col}{An excel column reference}
}
\description{
Convert excel column name to integer index
}
\examples{
convertFromExcelRef("DOG")
convertFromExcelRef("COW")

## numbers will be removed
convertFromExcelRef("R22")
}
\author{
Alexander Walker
}

