% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{copyWorkbook}
\alias{copyWorkbook}
\title{Copy a Workbook object.}
\usage{
copyWorkbook(wb)
}
\arguments{
\item{wb}{A workbook object}
}
\value{
Workbook
}
\description{
Just a wrapper of wb$copy()
}
\examples{

wb <- createWorkbook()  
wb2 <- wb ## does not create a copy
wb3 <- copyWorkbook(wb) ## wrapper for wb$copy()

addWorksheet(wb, "Sheet1") ## adds worksheet to both wb and wb2 but not wb3

names(wb)
names(wb2)
names(wb3)

}
