% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CommentClass.R
\name{createComment}
\alias{createComment}
\title{write a cell comment}
\usage{
createComment(comment, author = Sys.getenv("USERNAME"), style = NULL)
}
\arguments{
\item{comment}{Comment text. Character vector of length 1}

\item{author}{Author of comment. Character vector of length 1}

\item{wb}{A workbook object}

\item{sheet}{A vector of names or indices of worksheets}

\item{col}{Column a column number of letter}

\item{row}{A row number.}

\item{xy}{An alternative to specifying \code{col} and
\code{row} individually.  A vector of the form
\code{c(col, row)}.}
}
\description{
write a cell comment
}
\examples{
wb <- createWorkbook()
addWorksheet(wb, "Sheet 1")

writeComment(wb, 1, col = "A", row = 1,
   comment = "this is my comment")

saveWorkbook(wb, file = "writeCommentExample.xlsx", overwrite = TRUE)
}

