% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{dims_to_dataframe}
\alias{dims_to_dataframe}
\title{create dataframe from dimensions}
\usage{
dims_to_dataframe(dims, fill = FALSE)
}
\arguments{
\item{dims}{Character vector of expected dimension.}

\item{fill}{If TRUE, fills the dataframe with variables}
}
\description{
create dataframe from dimensions
}
\examples{
{
  dims_to_dataframe("A1:B2")
}
}
