% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getBaseFont}
\alias{getBaseFont}
\title{Return the workbook default font}
\usage{
getBaseFont(wb)
}
\arguments{
\item{wb}{A workbook object}
}
\description{
Return the workbook default font

Returns the base font used in the workbook.
}
\examples{
## create a workbook
wb <- createWorkbook()
getBaseFont(wb)

## modify base font to size 10 Arial Narrow in red
modifyBaseFont(wb, fontSize = 10, fontColour = "#FF0000", fontName = "Arial Narrow")

getBaseFont(wb)
}
\author{
Alexander Walker
}

