% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getNamedRegions}
\alias{getNamedRegions}
\title{Get named regions in an xlsx file.}
\usage{
getNamedRegions(xlsxFile)
}
\arguments{
\item{xlsxFile}{An xlsx file or Workbook object}
}
\description{
Get named regions in an xlsx file.
}
\examples{
##getNamedRegions
}

