% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{groupColumns}
\alias{groupColumns}
\title{Group columns}
\usage{
groupColumns(wb, sheet, cols, hidden = FALSE)
}
\arguments{
\item{wb}{A workbook object.}

\item{sheet}{A name or index of a worksheet.}

\item{cols}{Indices of cols to group.}

\item{hiddens}{Logical vector. If TRUE the grouped columns are hidden. Defaults to FALSE.}
}
\description{
Group a selection of columns
}
\details{
Group columns together, with the option to hide them.

NOTE: [setColWidths()] has a conflicting `hidden` parameter; changing one will reset the other.
}
\seealso{
[groupRows()] for grouping rows.
}
\author{
Joshua Sturm
}
