% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{groupRows}
\alias{groupRows}
\title{Group Rows}
\usage{
groupRows(wb, sheet, rows, hidden = FALSE)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{rows}{Indices of rows to group}

\item{hidden}{Logical vector. If TRUE the grouped columns are hidden. Defaults to false}
}
\description{
Group a selection of rows
}
\examples{
## Create a new workbook
wb <- createWorkbook()

## Add a worksheet
addWorksheet(wb, "Sheet 1") 

## Group first two rows
groupRows(wb, 1, rows = 1:2, hidden = FALSE)


## Save workbook
\dontrun{saveWorkbook(wb, "groupRowsExample.xlsx", overwrite = TRUE)}
}
\seealso{
\code{\link{groupColumns}}
}
\author{
Joshua Sturm
}
