% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{guess_col_type}
\alias{guess_col_type}
\title{function to estimate the column type.
0 = character, 1 = numeric, 2 = date.}
\usage{
guess_col_type(tt)
}
\arguments{
\item{tt}{dataframe produced by wb_to_df()}
}
\description{
function to estimate the column type.
0 = character, 1 = numeric, 2 = date.
}
