% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_null_then}
\alias{if_null_then}
\alias{\%||\%}
\title{If NULL then ...}
\usage{
x \%||\% y
}
\arguments{
\item{x}{A value to check}

\item{y}{A value to substitute if x is null}
}
\description{
Replace NULL
}
\examples{
\dontrun{
x <- NULL
x <- x \%||\% "none"
x <- x \%||\% NA
}

}
