% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{int2col}
\alias{int2col}
\title{Convert integer to Excel column}
\usage{
int2col(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Converts an integer to an Excel column label.
}
\examples{
int2col(1:10)
}

