% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{names}
\alias{names}
\alias{names.Workbook}
\alias{names<-.Workbook}
\title{get or set worksheet names}
\usage{
\method{names}{Workbook}(x)

\method{names}{Workbook}(x) <- value
}
\arguments{
\item{x}{A \code{Workbook} object}

\item{value}{a character vector the same length as wb}
}
\examples{

wb <- createWorkbook()
addWorksheet(wb, "S1")
addWorksheet(wb, "S2")
addWorksheet(wb, "S3")

names(wb)
names(wb)[[2]] <- "S2a"
names(wb)
names(wb) <- paste("Sheet", 1:3)
}

