% Generated by roxygen2 (4.0.1): do not edit by hand
\name{openXL}
\alias{openXL}
\title{Open an xlsx file}
\usage{
openXL(file=NULL)
}
\arguments{
\item{file}{path to the Excel (xls/xlsx) file}
}
\description{
This function tries to open an xlsx file with the
proper application, in a portable manner.
}
\examples{
example(writeData)
openXL("writeDataExample.xlsx")
}
\author{
Luca Braglia
}

