% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx.R
\name{openxlsx_options}
\alias{openxlsx_options}
\alias{op.openxlsx}
\alias{openxlsx_getOp}
\alias{openxlsx_setOp}
\title{openxlsx Options}
\usage{
op.openxlsx

openxlsx_getOp(x, default = NULL)

openxlsx_setOp(x, value)
}
\arguments{
\item{x}{An option name (\code{"openxlsx."} prefix optional)}

\item{default}{A default value if \code{NULL}}

\item{value}{The new value for the option (optional if x is a named list)}
}
\description{
See and get the openxlsx options
}
\details{
\code{openxlsx_getOp()} retrieves the \code{"openxlsx"} options found in
\code{op.openxlsx}.  If none are set (currently \code{NULL}) retrieves the
default option from \code{op.openxlsx}.  This will also check that the
intended option is a standard option (listed in \code{op.openxlsx}) and
will provide a warning otherwise.

\code{openxlsx_setOp()} is a safer way to set an option as it will first
check that the option is a standard option (as above) before setting.
}
\examples{
openxlsx_getOp("borders")
op.openxlsx[["openxlsx.borders"]]

}
