% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fontSizeLookupTables.R
\docType{data}
\name{openxlsxFontSizeLookupTable}
\alias{openxlsxFontSizeLookupTable}
\alias{openxlsxFontSizeLookupTableBold}
\title{Font Size Lookup tables}
\format{
A data.frame with column names corresponding to font names
}
\usage{
openxlsxFontSizeLookupTable

openxlsxFontSizeLookupTableBold
}
\description{
Lookup tables for font size
}
\examples{
data(openxlsxFontSizeLookupTable)
data(openxlsxFontSizeLookupTableBold)
}
\keyword{datasets}
