% Generated by roxygen2 (4.0.1): do not edit by hand
\name{orderCellRef}
\alias{orderCellRef}
\title{Returns ordering of sorted cell references}
\usage{
orderCellRef(x)
}
\arguments{
\item{x}{vector of cell references}
}
\value{
Ordering of sorted cell references
}
\description{
Returns ordering of sorted cell references
}

