% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{pageBreak}
\alias{pageBreak}
\title{add a page break to a worksheet}
\usage{
pageBreak(wb, sheet, i, type = "row")
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{i}{row or column number to insert page break.}

\item{type}{One of "row" or "column" for a row break or column break.}
}
\description{
add a page break to a worksheet
}
\examples{
wb <- createWorkbook()
addWorksheet(wb, "Sheet 1")
addWorksheet(wb, "Sheet 2")
addWorksheet(wb, "Sheet 3")

writeData(wb, 1, iris)
addFilter(wb, 1, row = 1, cols = 1:ncol(iris))

## Equivalently
writeData(wb, 2, x = iris, withFilter = TRUE)

## Similarly
writeDataTable(wb, 3, iris)

saveWorkbook(wb, file = "addFilterExample.xlsx", overwrite = TRUE)
}
\seealso{
\code{\link{addWorksheet}}
}

