% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{protectRange}
\alias{protectRange}
\title{Protect a range from modifications}
\usage{
protectRange(wb, sheet, range, password = NULL, name = NULL)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A worksheet where the range is located.}

\item{range}{A reference to the range to protect (example: \code{"A1:C3"}).}

\item{password}{An optional password to protect the range with.}

\item{name}{An optional name to give to the protected range. If none is given,
a random range name is created.}
}
\description{
Protect a range from modifications in the gui. Additionally, this allows a range to be unlocked
and edited separately from the worksheet when the worksheet is locked. This is
only fully enforced on a Windows version of Excel.
}
\details{
The password piece of this function does not work with Mac versions of Excel because there is no
concept of protected ranges there. That being said, you can create protected
ranges _without a password_, protect the worksheet, and then open the workbook
on a Mac, and you will be able to edit those ranges but not the rest of the 
sheet. Excel files with password protected ranges open without error on Mac, but
you will likely get an error if you try to edit them, rather than it asking
for a password.

Multiple ranges can be created with multiple calls to 
\code{protectRange()}.

While generally not useful, overlapping protected ranges are allowed
and are meaningful. Consider the case of protecting \code{"A1:C3"} with 
the password \code{"p1"}, and protecting \code{"A2:C3"} with the password
\code{"p2"}. The following situations could arise:

\itemize{
\item{}{Clicking in \code{"A1"} and entering \code{"p1"} will unlock all of \code{"A1:C3"}.}
\item{}{Clicking in \code{"A2"} and entering \code{"p1"} will unlock all of \code{"A1:C3"}.}
\item{}{Clicking in \code{"A2"} and entering \code{"p2"} will unlock all of \code{"A2:C3"}.
In this case, \code{"A1:C1"} will still be locked, and requires the password \code{"p1"} to unlock.} 
}
}
\examples{

# Create a workbook where range A1:C3 is password protected
# and the rest of the workbook is locked down and unaccessable
wb <- createWorkbook()
addWorksheet(wb, "S1")
protectRange(wb, "S1", range = "A1:C3", password = "password", name = "myrange")
protectWorksheet(wb, "S1")
saveWorkbook(wb, "Workbook_With_Range_Protection.xlsx")

# A password is not required, and if no name is supplied, 
# a random name is created 
wb <- createWorkbook()
addWorksheet(wb, "S1")
protectRange(wb, "S1", range = "A1:C3")
protectWorksheet(wb, "S1")
saveWorkbook(wb, "Workbook_With_Range_Protection_But_No_Password.xlsx")

# You can create a protected range without protecting the worksheet,
# but the protection will NOT be enforced until you protect the worksheet
# from within Excel
wb <- createWorkbook()
addWorksheet(wb, "S1")
protectRange(wb, "S1", range = "A1:C3", password = "password")
saveWorkbook(wb, "Workbook_With_Unactivated_Range_Protection.xlsx")

}
