% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{protectWorksheet}
\alias{protectWorksheet}
\title{Protect a worksheet from modifications}
\usage{
protectWorksheet(wb, sheet, protect = TRUE, password = NULL,
  selectLockedCells = NULL, selectUnlockedCells = NULL,
  formatCells = NULL, formatColumns = NULL, formatRows = NULL,
  insertColumns = NULL, insertRows = NULL, insertHyperlinks = NULL,
  deleteColumns = NULL, deleteRows = NULL, sort = NULL,
  autoFilter = NULL, pivotTables = NULL, objects = NULL,
  scenarios = NULL)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{protect}{Whether to protect or unprotect the sheet (default=TRUE)}

\item{password}{(optional) password required to unprotect the worksheet (not fully supported)}

\item{selectLockedCells}{Whether selecting locked cells is locked}

\item{selectUnlockedCells}{Whether selecting unlocked cells is locked}

\item{formatCells}{Whether formatting cells is locked}

\item{formatColumns}{Whether formatting columns is locked}

\item{formatRows}{Whether formatting rows is locked}

\item{insertColumns}{Whether inserting columns is locked}

\item{insertRows}{Whether inserting rows is locked}

\item{insertHyperlinks}{Whether inserting hyperlinks is locked}

\item{deleteColumns}{Whether deleting columns is locked}

\item{deleteRows}{Whether deleting rows is locked}

\item{sort}{Whether sorting is locked}

\item{autoFilter}{Whether auto-filter is locked}

\item{pivotTables}{Whether pivot tables are locked}

\item{objects}{Whether objects are locked}

\item{scenarios}{Whether scenarios are locked}
}
\description{
Protect or unprotect a worksheet from modifications by the user in the graphical user interface. Replaces an existing protection.
}
\examples{
wb <- createWorkbook()
addWorksheet(wb, "S1")
writeDataTable(wb, 1, x = iris[1:30,])
# Formatting cells / columns is allowed , but inserting / deleting columns is protected:
protectWorksheet(wb, "S1", protect = TRUE, formatCells = FALSE, formatColumns = FALSE, insertColumns = TRUE, deleteColumns = TRUE)

# Remove the protection
protectWorksheet(wb, "S1", protect = FALSE)

saveWorkbook(wb, "pageSetupExample.xlsx", overwrite = TRUE)
}
\author{
Reinhold Kainhofer
}
