% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentClass.R
\name{removeComment}
\alias{removeComment}
\title{Remove a comment from a cell}
\usage{
removeComment(wb, sheet, cols, rows, gridExpand = TRUE)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A vector of names or indices of worksheets}

\item{cols}{Columns to delete comments from}

\item{rows}{Rows to delete comments from}

\item{gridExpand}{If \code{TRUE}, all data in rectangle min(rows):max(rows) X min(cols):max(cols)
will be removed.}
}
\seealso{
\code{\link{createComment}}

\code{\link{writeComment}}
}

