% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{setActiveSheet}
\alias{setActiveSheet}
\title{Set the active worksheet of the workbook.}
\usage{
setActiveSheet(wb, ActiveSheet)
}
\arguments{
\item{wb}{A workbook object}

\item{ActiveSheet}{A string object with the name of the sheet or an integer 
with the ID of the sheet}
}
\description{
Just a wrapper of wb$setActiveSheet()
}
\examples{

wb <- createWorkbook()
addWorksheet(wb, "Sheet 1")
addWorksheet(wb, "Sheet 2")
addWorksheet(wb, "Sheet 3")
addWorksheet(wb, "Sheet 4")
setActiveSheet(wb, 2)

setActiveSheet(wb, "Sheet 3")
}
\author{
Philipp Schauberger
}
