% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sheetVisible}
\alias{sheetVisible}
\alias{sheetVisible<-}
\title{Get worksheet visible state.}
\usage{
sheetVisible(wb)

sheetVisible(wb) <- value
}
\arguments{
\item{wb}{A workbook object}

\item{value}{a logical vector the same length as sheetVisible(wb)}
}
\value{
Character vector of worksheet names.

TRUE if sheet is visible, FALSE if sheet is hidden
}
\examples{

wb <- createWorkbook()
addWorksheet(wb, sheetName = "S1", visible = FALSE)
addWorksheet(wb, sheetName = "S2", visible = TRUE)
addWorksheet(wb, sheetName = "S3", visible = FALSE)

sheetVisible(wb)
sheetVisible(wb)[1] <- TRUE ## show sheet 1
sheetVisible(wb)[2] <- FALSE ## hide sheet 2

}
\author{
Alexander Walker
}

