% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_xlsx}
\alias{temp_xlsx}
\title{helper function to create tempory directory for testing purpose}
\usage{
temp_xlsx(name = "temp_xlsx")
}
\description{
helper function to create tempory directory for testing purpose
}
