% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ungroupColumns}
\alias{ungroupColumns}
\title{Ungroup Columns}
\usage{
ungroupColumns(wb, sheet, cols)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{cols}{Indices of columns to ungroup}
}
\description{
Ungroup a selection of columns
}
\details{
If column was previously hidden, it will now be shown
}
\seealso{
\code{\link{ungroupRows}} To ungroup rows
}
\author{
Joshua Sturm
}
