% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ungroupRows}
\alias{ungroupRows}
\title{Ungroup Rows}
\usage{
ungroupRows(wb, sheet, rows)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{rows}{Indices of rows to ungroup}
}
\description{
Ungroup a selection of rows
}
\details{
If row was previously hidden, it will now be shown
}
\seealso{
\code{\link{ungroupColumns}}
}
\author{
Joshua Sturm
}
