% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cell.R
\name{update_cell}
\alias{update_cell}
\title{replace data in a single cell}
\usage{
update_cell(x, wb, sheet, cell)
}
\arguments{
\item{x}{value you want to insert}

\item{wb}{the workbook you want to update}

\item{sheet}{the sheet you want to update}

\item{cell}{the cell you want to update in Excel conotation e.g. "A1"}
}
\description{
Minimal invasive update of a single cell for imported workbooks.
}
\examples{
   xlsxFile <- system.file("extdata", "update_test.xlsx", package = "openxlsx")
   wb <- loadWorkbook(xlsxFile)
   wb <- update_cell(x = c(1:3), wb = wb, sheet = "Sheet1", cell = "D4:D6")
   wb <- update_cell(x = 7, wb = wb, sheet = "Sheet1", cell = "H1")
}
