% Generated by roxygen2 (4.0.2): do not edit by hand
\name{view}
\alias{view}
\title{Alternative to View that uses openXL}
\usage{
view(...)
}
\arguments{
\item{...}{writeData handled objects}
}
\description{
Alternative to \code{\link{View}} that invokes a spreasheet viewer on
openxlsx's \code{\link{writeData}} handled objects.
}
\examples{
\dontrun{
view(Indometh, iris)
}
}

