% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData2.R
\name{writeData2}
\alias{writeData2}
\title{dummy function to write data}
\usage{
writeData2(
  wb,
  sheet,
  data,
  colNames = TRUE,
  rowNames = FALSE,
  startRow = 1,
  startCol = 1
)
}
\arguments{
\item{wb}{workbook}

\item{sheet}{sheet}

\item{data}{data to export}

\item{colNames}{include colnames?}

\item{rowNames}{include rownames?}

\item{startRow}{row to place it}

\item{startCol}{col to place it}
}
\description{
dummy function to write data
}
\examples{
# create a workbook and add some sheets
wb <- createWorkbook()

addWorksheet(wb, "sheet1")
writeData2(wb, "sheet1", mtcars, colNames = TRUE, rowNames = TRUE)

addWorksheet(wb, "sheet2")
writeData2(wb, "sheet2", cars, colNames = FALSE)

addWorksheet(wb, "sheet3")
writeData2(wb, "sheet3", letters)

addWorksheet(wb, "sheet4")
writeData2(wb, "sheet4", as.data.frame(Titanic), startRow = 2, startCol = 2)

saveWorkbook(wb, file = "/tmp/test.xlsx", overwrite = TRUE)

}
