% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{all.equal}
\alias{all.equal}
\alias{all.equal.Workbook}
\title{Check equality of workbooks}
\usage{
\method{all.equal}{Workbook}(target, current, ...)
}
\arguments{
\item{target}{A \code{Workbook} object}

\item{current}{A \code{Workbook} object}

\item{...}{ignored}
}

