% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{as.character.formula}
\alias{as.character.formula}
\title{\code{as.character.formula()}}
\usage{
\method{as.character}{formula}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{Not implemented}
}
\value{
\code{base::as.character.default(x)}
}
\description{
This function exists to prevent conflicts with \code{as.character.formula} methods
from other packages
}
