% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chooseExcelApp}
\alias{chooseExcelApp}
\title{Search for available xls/xlsx reader application in a
UNIX like (mainly Linux) environment.}
\usage{
chooseExcelApp()
}
\description{
This function search for available xls/xlsx
reader application. If it founds anything, sets
\code{options('openxlsx.excel.app')}, used by \code{openXL}.
}
\author{
Luca Braglia
}

