% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{conditionalFormat}
\alias{conditionalFormat}
\title{Add conditional formatting to cells}
\usage{
conditionalFormat(wb, sheet, cols, rows, rule = NULL, style = NULL,
  type = "expression")
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{cols}{Columns to apply conditional formatting to}

\item{rows}{Rows to apply conditional formatting to}

\item{rule}{The condition under which to apply the formatting or a vector of colours. See examples.}

\item{style}{A style to apply to those cells that satisify the rule. A Style object returned from createStyle()}

\item{type}{Either 'expression', 'colorscale' or 'databar'. If 'expression' the formatting is determined
by a formula.  If colorScale cells are coloured based on cell value. See examples.}
}
\description{
DEPRECATED! USE \code{\link{conditionalFormatting}}
}
\details{
DEPRECATED! USE \code{\link{conditionalFormatting}}

Valid operators are "<", "<=", ">", ">=", "==", "!=". See Examples.
Default style given by: createStyle(fontColour = "#9C0006", bgFill = "#FFC7CE")
}
\seealso{
\code{\link{createStyle}}
}
\author{
Alexander Walker
}
