% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{convertFromExcelRef}
\alias{convertFromExcelRef}
\title{Convert excel column name to integer index}
\usage{
convertFromExcelRef(col)
}
\arguments{
\item{col}{An excel column reference}
}
\description{
Convert excel column name to integer index e.g. "J" to 10
}
\examples{
convertFromExcelRef("DOG")
convertFromExcelRef("COW")

## numbers will be removed
convertFromExcelRef("R22")
}
