% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{convertToDate}
\alias{convertToDate}
\title{Convert from excel date number to R Date type}
\usage{
convertToDate(x, origin = "1900-01-01", ...)
}
\arguments{
\item{x}{A vector of integers}

\item{origin}{date. Default value is for Windows Excel 2010}

\item{...}{additional parameters passed to as.Date()}
}
\details{
Excel stores dates as number of days from some origin day
}
\examples{
##2014 April 21st to 25th
convertToDate(c(41750, 41751, 41752, 41753, 41754, NA) )
convertToDate(c(41750.2, 41751.99, NA, 41753 ))
}
\seealso{
\code{\link{writeData}}
}

