% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{createWorkbook}
\alias{createWorkbook}
\title{Create a new Workbook object}
\usage{
createWorkbook(creator = Sys.getenv("USERNAME"))
}
\arguments{
\item{creator}{Creator of the workbook (your name). Defaults to login username}
}
\value{
Workbook object
}
\examples{
## Create a new workbook
wb <- createWorkbook()

## Save workbook to working directory
saveWorkbook(wb, file = "createWorkbookExample", overwrite = TRUE)
}
\author{
Alexander Walker
}
\seealso{
\code{\link{loadWorkbook}}

\code{\link{saveWorkbook}}
}

