% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getCellRefs}
\alias{getCellRefs}
\title{Return excel cell coordinates from (x,y) coordinates}
\usage{
getCellRefs(cellCoords)
}
\arguments{
\item{cellCoords}{A data.frame with two columns coordinate pairs.}
}
\value{
Excel alphanumeric cell reference
}
\author{
Alexander Walker
}

