% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getCreators}
\alias{getCreators}
\title{Add another author to the meta data of the file.}
\usage{
getCreators(wb)
}
\arguments{
\item{wb}{A workbook object}
}
\value{
vector of creators
}
\description{
Just a wrapper of wb$getCreators()
Get the names of the
}
\examples{

wb <- createWorkbook()
getCreators(wb)
}
\author{
Philipp Schauberger
}
