% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getStyles}
\alias{getStyles}
\title{Returns a list of all styles in the workbook}
\usage{
getStyles(wb)
}
\arguments{
\item{wb}{A workbook object}
}
\examples{
## load a workbook 
wb <- loadWorkbook(file = system.file("loadExample.xlsx", package = "openxlsx"))
getStyles(wb)[1:3]
}
\author{
Alexander Walker
}
\seealso{
\code{\link{replaceStyle}}
}

