% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{groupRows}
\alias{groupRows}
\title{Group Rows}
\usage{
groupRows(wb, sheet, rows, hidden = FALSE)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{rows}{Indices of rows to group}

\item{hidden}{Logical vector. If TRUE the grouped columns are hidden. Defaults to FALSE}
}
\description{
Group a selection of rows
}
\seealso{
\code{\link{ungroupRows}} to ungroup rows. \code{\link{groupColumns}} for grouping columns.
}
\author{
Joshua Sturm
}
