% Generated by roxygen2 (4.0.1): do not edit by hand
\name{openXL}
\alias{openXL}
\title{Open a Microsoft Excel file (xls/xlsx) or an openxlsx Workbook}
\usage{
openXL(file=NULL)
}
\arguments{
\item{file}{path to the Excel (xls/xlsx) file or Workbook object.}
}
\description{
This function tries to open a Microsoft Excel
(xls/xlsx) file or an openxlsx Workbook with the proper
application, in a portable manner.

In Linux it searches for available xls/xlsx reader application
(unless \code{options('openxlsx.excelApp')} is set to the bin
app path), and if it founds anything, sets
\code{options('openxlsx.excelApp')}.
}
\examples{
# file example
example(writeData)
openXL("writeDataExample.xlsx")

# (not yet saved) Workbook example
wb <- createWorkbook()
x <- mtcars[1:6,]
addWorksheet(wb, "Cars")
writeData(wb, "Cars", x, startCol = 2, startRow = 3, rowNames = TRUE)
openXL(wb)
}
\author{
Luca Braglia
}

