% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{openxlsx}
\alias{openxlsx}
\alias{openxlsx-package}
\title{xlsx reading, writing and editing.}
\description{
Roxygen is a Doxygen-like documentation system for R; allowing
in-source specification of Rd files, collation and namespace
directives.
}
\details{
The openxlsx package uses global options to simplfy formatting:

\itemize{
   \item{\code{options("openxlsx.borderColour" = "black")}}
   \item{\code{options("openxlsx.borderStyle" = "thin")}}
   \item{\code{ooptions("openxlsx.dateFormat" = "mm/dd/yyyy")}}
   \item{\code{options("openxlsx.dateFormat" = "mm/dd/yyyy")}}
  }
}
\author{
Alex Walker \email{Alexander.Walker@gmail.com},
}
\seealso{
See \code{vignette("examples", package = "openxlsx")} and
 \code{vignette("formatting", package = "openxlsx")} for examples
}

