% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx.R
\docType{data}
\name{openxlsx_options}
\alias{openxlsx_options}
\alias{openxlsx_getOp}
\alias{op.openxlsx}
\title{openxlsx Options}
\format{
An object of class \code{list} of length 16.
}
\usage{
openxlsx_getOp(x)

op.openxlsx
}
\arguments{
\item{x}{An option name (\code{"openxlsx."} prefix optional)}
}
\description{
See and get the openxlsx options
}
\details{
\code{openxlsx_getOp()} retrieves the \code{"openxlsx"} options found in
  \code{op.openxlsx}.  If none are set (currently `NULL`) retrieves the 
  default option from \code{op.openxlsx}
}
\examples{
openxlsx_getOp("borders")
op.openxlsx[["openxlsx.borders"]]

}
\keyword{datasets}
