% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{protectRange}
\alias{protectRange}
\title{Protect a range from modifications}
\usage{
protectRange(wb, sheet, range, password = NULL, name = NULL)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A worksheet where the range is located}

\item{range}{A reference to the range to protect (example: `"A1:C3"`)}

\item{password}{A password to protect the range with}

\item{name}{A name to give to the protected range}
}
\description{
Protect a range from modifications in the gui. Additionally, this allows a range to be unlocked
and edited separately from the worksheet when the worksheet is locked.
}
\examples{

wb <- createWorkbook()
addWorksheet(wb, "S1")
protectRange(wb, "S1", range = "A1:C3", password = "password", name = "myrange")
protectWorksheet(wb, "sheet")
saveWorkbook(wb, "Workbook_With_Range_Protection.xlsx")

}
