% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{setLastModifiedBy}
\alias{setLastModifiedBy}
\title{Add another author to the meta data of the file.}
\usage{
setLastModifiedBy(wb, LastModifiedBy)
}
\arguments{
\item{wb}{A workbook object}

\item{LastModifiedBy}{A string object with the name of the LastModifiedBy-User}
}
\description{
Just a wrapper of wb$changeLastModifiedBy()
}
\examples{

wb <- createWorkbook()
setLastModifiedBy(wb, "test")
}
\author{
Philipp Schauberger
}
