% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sheets}
\alias{sheets}
\title{Returns names of worksheets.}
\usage{
sheets(wb)
}
\arguments{
\item{wb}{A workbook object}
}
\value{
Name of worksheet(s) for a given index
}
\details{
DEPRECATED. Use \code{\link{names}}
}
\examples{

## Create a new workbook
wb <- createWorkbook()

## Add some worksheets
addWorksheet(wb, "Worksheet Name")
addWorksheet(wb, "This is worksheet 2")
addWorksheet(wb, "The third worksheet")

## Return names of sheets, can not be used for assignment.
names(wb)
# openXL(wb)

names(wb) <- c("A", "B", "C")
names(wb)
# openXL(wb)

}
\author{
Alexander Walker
}
\seealso{
\code{\link{names}} to rename a worksheet in a Workbook
}

