% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cell.R
\name{update_cell}
\alias{update_cell}
\title{replace data in a single cell}
\usage{
update_cell(x, wb, sheet, cell, data_class, colNames = FALSE)
}
\arguments{
\item{x}{value you want to insert}

\item{wb}{the workbook you want to update}

\item{sheet}{the sheet you want to update}

\item{cell}{the cell you want to update in Excel conotation e.g. "A1"}

\item{data_class}{optional data class object}

\item{colNames}{if TRUE colNames are passed down}
}
\description{
Minimal invasive update of a single cell for imported workbooks.
}
\examples{
   xlsxFile <- system.file("extdata", "update_test.xlsx", package = "openxlsx")
   wb <- loadWorkbook(xlsxFile)
   wb <- update_cell(x = c(1:3), wb = wb, sheet = "Sheet1", cell = "D4:D6")
   wb <- update_cell(x = c("x", "y", "z"), wb = wb, sheet = "Sheet1", cell = "B3:D3")
   wb <- update_cell(x = 7, wb = wb, sheet = "Sheet1", cell = "D4")
   
   wb <- loadWorkbook(xlsxFile)
   wb <- update_cell(x = 7, wb = wb, sheet = "Sheet1", cell = "A4")
   
   wb_to_df(wb)
   
}
