% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_functions.R
\name{wb_to_df}
\alias{wb_to_df}
\title{Create Dataframe from Workbook}
\usage{
wb_to_df(
  wb,
  sheet,
  colNames = TRUE,
  dims,
  detectDates = TRUE,
  showFormula = FALSE,
  convert = TRUE,
  skipEmptyCols = FALSE
)
}
\arguments{
\item{wb}{openxlsx workbook}

\item{sheet}{Either sheet name or index. When missing the first sheet in the workbook is selected.}

\item{dims}{Character string of type "A1:B2" as optional dimentions to be imported.}

\item{detectDates}{If TRUE, attempt to recognise dates and perform conversion.}

\item{showFormula}{If TRUE, the underlying Excel formulas are shown.}

\item{convert}{If TRUE, a conversion to dates and numerics is attempted.}

\item{skipEmptyCols}{If TRUE, empty columns are skipped.}

\item{colnames}{If TRUE, the first row of data will be used as column names.}
}
\description{
Simple function to create a dataframe from a workbook. Simple as in simply
written down and not optimized etc. The goal was to have something working.
}
\examples{
  # numerics, dates, missings, bool and string
  xlsxFile <- system.file("extdata", "readTest.xlsx", package = "openxlsx")
  wb1 <- loadWorkbook(xlsxFile)
  
  # inlinestr
  xlsxFile <- system.file("extdata", "inlinestr.xlsx", package = "openxlsx")
  wb2 <- loadWorkbook(xlsxFile)

  wb_to_df(wb1)
  wb_to_df(wb1, colNames = FALSE)
  wb_to_df(wb1, detectDates = FALSE)
  wb_to_df(wb1, showFormula = TRUE)
  wb_to_df(wb1, dims = "A2:C5", colNames = FALSE)
  wb_to_df(wb1, convert = FALSE)
  wb_to_df(wb1, skipEmptyCols = TRUE)
  read.xlsx(wb1)

  wb_to_df(wb2)
  read.xlsx(wb2)

}
