% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeData}
\alias{writeData}
\title{Write an object to a worksheet}
\usage{
writeData(wb = NULL, sheet = NULL, x = NULL, startCol = 1,
  startRow = 1, xy = NULL, colNames = TRUE, rowNames = FALSE,
  headerStyle = NULL, borders = c("none", "surrounding", "rows", "columns"),
  borderColour = getOption("openxlsx.borderColour", "black"),
  borderStyle = getOption("openxlsx.borderStyle", "thin"), gridLines = TRUE,
  ...)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The worksheet to write to. Can be the worksheet
index or name. If \code{NULL} (default) it's inferred from
\code{x} name. It can be a numeric index. If not present in
wb, it will be created via addWorksheet.}

\item{x}{Object to be written. For classes supported look at the examples.}

\item{startCol}{A vector specifiying the starting columns(s) to write to.}

\item{startRow}{A vector specifiying the starting row(s) to write to.}

\item{xy}{An alternative to specifying \code{startCol} and
\code{startRow} individually.  A vector of the form
\code{c(startCol, startRow)}.}

\item{colNames}{If \code{TRUE}, column names of x are written.}

\item{rowNames}{If \code{TRUE}, data.frame row names of x are written.}

\item{headerStyle}{Custom style to apply to column names.}

\item{borders}{Either "\code{none}" (default), "\code{surrounding}",
"\code{columns}", "\code{rows}" or \emph{respective abbreviations}.  If
"\code{surrounding}", a border is drawn around the data.  If "\code{rows}",
a surrounding border is drawn with a border around each row. If
"\code{columns}", a surrounding border is drawn with a border between
each column.}

\item{borderColour}{Colour of cell border.  A valid colour (belonging to \code{colours()} or a hex colour code, eg see \href{http://www.colorpicker.com}{here}).}

\item{borderStyle}{Border line style
\itemize{
   \item{\bold{none}}{ no border}
   \item{\bold{thin}}{ thin border}
   \item{\bold{medium}}{ medium border}
   \item{\bold{dashed}}{ dashed border}
   \item{\bold{dotted}}{ dotted border}
   \item{\bold{thick}}{ thick border}
   \item{\bold{double}}{ double line border}
   \item{\bold{hair}}{ hairline border}
   \item{\bold{mediumDashed}}{ medium weight dashed border}
   \item{\bold{dashDot}}{ dash-dot border}
   \item{\bold{mediumDashDot}}{ medium weight dash-dot border}
   \item{\bold{dashDotDot}}{ dash-dot-dot border}
   \item{\bold{mediumDashDotDot}}{ medium weight dash-dot-dot border}
   \item{\bold{slantDashDot}}{ slanted dash-dot border}
  }}

\item{gridLines}{gridLines passed to addWorksheet if a new
worksheet is neede}

\item{...}{Further arguments (for future use)}
}
\description{
Write an object to worksheet with optional styling.
}
\examples{
wb <- createWorkbook()
options("openxlsx.borderStyle" = "thin")
options("openxlsx.borderColour" = "#4F81BD")
## Add worksheets
addWorksheet(wb, "Cars")


x <- mtcars[1:6,]
writeData(wb, "Cars", x, startCol = 2, startRow = 3, rowNames = TRUE)
writeData(wb, "Cars", x, rowNames = TRUE, startCol = "O", startRow = 3,
         borders="surrounding", borderColour = "black") ## black border

writeData(wb, "Cars", x, rowNames = TRUE,
         startCol = 2, startRow = 12, borders="columns")

writeData(wb, "Cars", x, rowNames = TRUE,
         startCol="O", startRow = 12, borders="rows")

## header styles
hs1 <- createStyle(fgFill = "#DCE6F1", halign = "CENTER", textDecoration = "Italic",
                   border = "Bottom")

writeData(wb, "Cars", x, colNames = TRUE, rowNames = TRUE, startCol="B",
     startRow = 23, borders="rows", headerStyle = hs1, borderStyle = "dashed")

hs2 <- createStyle(fontColour = "#ffffff", fgFill = "#4F80BD",
                   halign = "center", valign = "center", textDecoration = "Bold",
                   border = "TopBottomLeftRight")

writeData(wb, "Cars", x, colNames = TRUE, rowNames = TRUE,
          startCol="O", startRow = 23, borders="columns", headerStyle = hs2)

## Save workbook
saveWorkbook(wb, "writeDataExample.xlsx", overwrite = TRUE)

\dontrun{

## inspired by xtable gallery
##' http://cran.r-project.org/web/packages/xtable/vignettes/xtableGallery.pdf

## Create a new workbook
wb <- createWorkbook()
data(tli, package = "xtable")

## TEST 1 - data.frame
test.n <- "data.frame"
my.df <- tli[1:10, ]
writeData(wb = wb, sheet = test.n, x = my.df, borders = "n")

## TEST 2 - matrix
test.n <- "matrix"
design.matrix <- model.matrix(~ sex * grade, data = my.df)
writeData(wb = wb, sheet = test.n, x = design.matrix)

## TEST 3 - aov
test.n <- "aov"
fm1 <- aov(tlimth ~ sex + ethnicty + grade + disadvg, data = tli)
writeData(wb = wb, sheet = test.n, x = fm1)

## TEST 4 - lm
test.n <- "lm"
fm2 <- lm(tlimth ~ sex*ethnicty, data = tli)
writeData(wb = wb, sheet = test.n, x = fm2)

## TEST 5 - anova 1
test.n <- "anova"
my.anova <- anova(fm2)
writeData(wb = wb, sheet = test.n, x = my.anova)

## TEST 6 - anova 2
test.n <- "anova2"
fm2b <- lm(tlimth ~ ethnicty, data = tli)
my.anova2 <- anova(fm2b, fm2)
writeData(wb = wb, sheet = test.n, x = my.anova2)

## TEST 7 - GLM
test.n <- "glm"
fm3 <- glm(disadvg ~ ethnicty*grade, data = tli, family = binomial())
writeData(wb = wb, sheet = test.n, x = fm3)

## TEST 8 - prcomp
test.n <- "prcomp"
pr1 <- prcomp(USArrests)
writeData(wb = wb, sheet = test.n, x = pr1)

## TEST 9 - summary.prcomp
test.n <- "summary.prcomp"
writeData(wb = wb, sheet = test.n, x = summary(pr1))

## TEST 10 - simple table
test.n <- "table"
data(airquality)
airquality$OzoneG80 <- factor(airquality$Ozone > 80,
                              levels = c(FALSE, TRUE),
                              labels = c("Oz <= 80", "Oz > 80"))
airquality$Month <- factor(airquality$Month,
                           levels = 5:9,
                           labels = month.abb[5:9])
my.table <- with(airquality, table(OzoneG80,Month) )
writeData(wb = wb, sheet = test.n, x = my.table)

## Save workbook
saveWorkbook(wb, "classTests.xlsx",  overwrite = TRUE)
}
}
\author{
Alexander Walker
}
\seealso{
\code{\link{writeData}}
}

